Always display all markers on a given zlevel, add triangle (dialog arrow) to status-button-menu. Move some methods to separate file to tidy up code.
