+++ src/app/status-button-menu/status-button-menu.component.scss
@@ -3,6 +3,43 @@
     width: 100%;
 }
 
+#statusSmileyButtonMenu:after {
+    content: " ";
+    position: absolute;
+    right: 10px;
+    bottom: -15px;
+    border-top: 15px solid #fff;
+    border-right: 15px solid transparent;
+    border-left: 15px solid transparent;
+    border-bottom: none;
+  }
+
+@media only screen and (min-width: 720px) {
+    #statusSmileyButtonMenu:after {
+        content: " ";
+        position: absolute;
+        right: 20px;
+        bottom: -15px;
+        border-top: 15px solid #fff;
+        border-right: 15px solid transparent;
+        border-left: 15px solid transparent;
+        border-bottom: none;
+      }
+}
+
+@media only screen and (min-width: 1024px) and (min-height: 1024px) {
+    #statusSmileyButtonMenu:after {
+        content: " ";
+        position: absolute;
+        right: 35px;
+        bottom: -15px;
+        border-top: 15px solid #fff;
+        border-right: 15px solid transparent;
+        border-left: 15px solid transparent;
+        border-bottom: none;
+      }
+}
+
 #statusSmileyButtonMenu nb-card-body {
      padding-top: 0px !important;
      padding-bottom: 0px !important;
